webapi/webapi-vehicleinfo-ivi-tests/vehicleinfo/Zone_attributes_tests.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Liu,xin <xinx.liu@intel.com>

-->

<meta charset="utf-8">
<title>Vehicle Test: Vehicle_attributes</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://rawgit.com/w3c/automotive-bg/master/vehicle_spec.html">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="support/support.js"></script>
<div id="log"></div>
<script>
var zone;
test(function () {
  zone = new Zone;
}, "Check that Zone is a constructor");

test(function () {
  assert_equals(typeof zone, "object");
}, "Check that Zone returns an object");

test(function () {
  assert_true("value" in zone);
}, "Check if Zone.value exists");

test(function () {
  assert_equals(typeof zone.value, "object");
}, "Check that Zone.value returns an object array");

test(function () {
  assert_true("driver" in zone);
}, "Check if Zone.driver exists");

test(function () {
  assert_equals(typeof zone.driver, "object");
}, "Check that Zone.driver returns an object Zone");

test(function () {
  assert_true("equals" in zone);
}, "Check if Zone.equals exists");

test(function() {
  var zone1 = new Zone(["Front", "Left"])
  assert_equals(typeof zone.equals(zone1), "boolean");
}, "Check that Zone.equals returns a boolean");

test(function () {
  assert_true("contains" in zone);
}, "Check if Zone.contains exists");

test(function () {
  var zone2 = new Zone(["Front", "Right"])
  assert_equals(typeof zone.contains(zone2), "boolean");
}, "Check that Zone.contains returns a boolean");

</script>
